Export_Software/requirements.txt
PySide6>=6.8,<7
pyserial>=3.5,<4
qtawesome>=1.3,<2
